requirements.txt
pillow
keyboard
baidu-aip
pywin32
win10toast